requirements.txt
django ==1.9.10
markdown
Pygments
django-summernote
xadmin
django-dashing
django-ckeditor
whoosh
django-haystack
jieba
django-imagekit
celery
django-crispy-forms==1.6.1
#shop page
django-el-pagination
#restful
djangorestframework==3.6.4

